Changes to Source for Java v19
Netbeans language suggestions & configuration for strong
encapsulation exceptions required by Java 8 -> Java 9 to
enable certain functionality like File->New. It is much easier
to add exceptions instead of fixing because the deep reflection
calls are inside libraries.

+++ Drawables/src/com/billooms/drawables/vecmath/Vector3d.java
@@ -171,7 +171,7 @@ public class Vector3d extends Tuple3d implements java.io.Serializable {
     if (vDot > 1.0) {
       vDot = 1.0;
     }
-    return ((double) (Math.acos(vDot)));
+    return Math.acos(vDot);
   }
 
 }
